+++ Servidor/IBhaskara.java
@@ -2,5 +2,5 @@ import java.rmi.Remote;
 import java.rmi.RemoteException;
 
 public interface IBhaskara extends Remote{
-  String calcBhaskara(int a, int b, int c) throws RemoteException;
+  String calcBhaskara(double a, double b, double c) throws RemoteException;
 }
